--- tfts/models/tasks/prediction.py
@@ -1,6 +1,0 @@
-import tensorflow as tf
-
-
-class PredictionHead(tf.keras.layers.Layer):
-    def __init__(self):
-        super(PredictionHead, self).__init__()
